\name{bootRes-internal}

%% <LIST>
\alias{bcf}
\alias{brf}
\alias{pmat}
%% </LIST>

\title{Internal bootRes Functions}
\description{
  Internal bootRes functions
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}
\keyword{ internal }