\name{bootRes-package}
\Rdversion{1.1}
\alias{bootRes-package}
\alias{bootRes}
\docType{package}
\title{
The bootRes Package for Bootstrapped Response and Correlation Functions
}
\description{
This package contains functions for calculating bootstrapped response and correlation functions for use in dendroclimatology.
}
\details{
\tabular{ll}{
Package: \tab bootRes\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2009-06-24\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
Function \code{\link{dendroclim}} is used to calculate response and correlation functions using tree-ring chronologies and monthly climatic data (Fritts, 1976, Cook and Kairiukstis, 1990), \code{\link{dcplot}} is a convenient plotting function for its output. In its current state this package is a clone of the computer programme DENDROCLIM2002 (Biondi and Waikul, 2004).
}
\author{
Christian Zang

Maintainer: Christian Zang <christian.zang@gmail.com>
}

\references{
	Biondi, F. & Waikul, K. (2004) DENDROCLIM2002: A C++ program for statistical calibration of climate signals in tree-ring chronologies. \emph{Computers & Geosciences} 30:303-311
	
	Cook, E.R. and Kairiukstis, L.A. (1990) \emph{Methods of Dendrochronology: Applications in the Environmental Sciences.} Springer
	
	Fritts, H.C. (1976) \emph{Tree ring and climate}. Academic Press
}

\keyword{ package }
