\name{muc.clim}
\docType{data}
\alias{muc.clim}
\title{Monthly Mean Temperature and Total Precipitation for Forstenrieder Park, Munich}
\description{
  This dataset gives the monthly mean temperature and total precipitation at Forstenrieder Park, Munich, Bavaria, Germany. 
}
\usage{muc.clim}
\format{A \code{data.frame} containing four columns with year, month, temperature and precipitation. }
\references{
  Kagerer, K. (2009) Wachstum von Fichte, Kiefer und Eiche in Abhaengigkeit von Klima und Kronenklasse im Forstenrieder Park, Muenchen. Diploma-Thesis, University of Applied Sciences Weihenstephan.
}
\keyword{datasets}