\name{muc.fake}
\docType{data}
\alias{muc.fake}
\title{Modeled Tree-Ring Chronology of a Spruce Population near Munich}
\description{
  This dataset gives the modelled tree-ring widths for \emph{Picea abies}
  at Forstenrieder Park, Munich, Bavaria, Germany. Tree growth was modeled as a response of low temperatures in previous and current July and August, high temperatures in current February and March, and high precipitation amounts in current July and August.
}
\usage{muc.fake}
\format{A \code{data.frame} containing tree-ring indices with respective years as \code{rownames}.
  in rows. }
\references{
  Kagerer, K. (2009) Wachstum von Fichte, Kiefer und Eiche in Abhaengigkeit von Klima und Kronenklasse im Forstenrieder Park, Muenchen. Diploma-Thesis, University of Applied Sciences Weihenstephan.
}
\keyword{datasets}