\name{muc.spruce}
\docType{data}
\alias{muc.spruce}
\title{Tree-Ring Chronology of a Spruce Population near Munich}
\description{
  This dataset gives the tree-ring indices for \emph{Picea abies}
  at Forstenrieder Park, Munich, Bavaria, Germany. The chronology represents 20 trees.
}
\usage{muc.spruce}
\format{A \code{data.frame} containing tree-ring indices with respective years as \code{rownames}.
  in rows. }
\references{
  Kagerer, K. (2009) Wachstum von Fichte, Kiefer und Eiche in Abhaengigkeit von Klima und Kronenklasse im Forstenrieder Park, Muenchen. Diploma-Thesis, University of Applied Sciences Weihenstephan.
}
\keyword{datasets}