\name{rt.prec}
\docType{data}
\alias{rt.prec}
\title{Monthly Precipitation Sums for Rothenburg ob der Tauber, Germany}
\description{
  This dataset gives the monthly precipitation sum  at Rothenburg ob der
  Tauber, Bavaria, Germany in a decadal (DENDROCLIM2002-style) format)
}
\usage{rt.prec}
\format{A \code{data.frame} containing thirteen columns with year and
  twelve months of precipitation data in degree Celsius. }
\references{
  Zang, C. (2010) Growth reaction of precerate forest trees to summer
  drought -- a multispecies tree-ring network approach. Ph.D. Thesis,
  Technische Universitaet Muenchen, 164 pp.
}
\keyword{datasets}