\name{rt.temp}
\docType{data}
\alias{rt.temp}
\title{Monthly Mean Temperature for Rothenburg ob der Tauber, Germany}
\description{
  This dataset gives the monthly mean temperature at Rothenburg ob der
  Tauber, Bavaria, Germany in a decadal (DENDROCLIM2002-style) format)
}
\usage{rt.temp}
\format{A \code{data.frame} containing thirteen columns with year and
  twelve months of temperature data in degree Celsius. }
\references{
  Zang, C. (2010) Growth reaction of temperate forest trees to summer
  drought -- a multispecies tree-ring network approach. Ph.D. Thesis,
  Technische Universitaet Muenchen, 164 pp.
}
\keyword{datasets}