\name{rt.spruce}
\docType{data}
\alias{rt.spruce}
\title{Tree-Ring Chronology of a Spruce Population at Rothenburg ob der Tauber}
\description{
  This dataset gives the tree-ring indices for \emph{Picea abies}
  at Rothenburg ob der Tauber, Bavaria, Germany. The chronology represents 20 trees.
}
\usage{rt.spruce}
\format{A \code{data.frame} containing tree-ring indices with respective years as \code{rownames}.
  in rows. }
\references{
  Zang, C. (2010) Growth reaction of precerate forest trees to summer
  drought -- a multispecies tree-ring network approach. Ph.D. Thesis,
  Technische Universitaet Muenchen, 164 pp.
}
\keyword{datasets}